+++ app/src/main/java/com/example/unicourse/ui/fragments/ProfileFragment.java
@@ -148,7 +148,7 @@ public class ProfileFragment extends Fragment {
                             int accomplishAmountCount = 0;
                             if (progressResponse.getData() != null) {
                                 for (EnrolledCourseProgressResponse.EnrolledCourse course : progressResponse.getData()) {
-                                    profileCourses.add(new ProfileCourse(course.getCourse().getTitle(), course.getCourse().getThumbnail()));
+                                    profileCourses.add(new ProfileCourse(course.getCourse().get_id(), course.getCourse().getTitle(), course.getCourse().getThumbnail()));
                                     progressAmountCount += course.getProgress();
                                     if (course.isCompleted()) {
                                         accomplishAmountCount++;
