+++ app/src/main/res/layout/activity_cart_course_item.xml
@@ -39,6 +39,7 @@
                 android:layout_height="13dp"
                 android:src="@drawable/activity_cart_checkbox_ic_checked"
                 app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="@+id/cartCheckBox"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent" />
         </androidx.constraintlayout.widget.ConstraintLayout>
@@ -119,6 +120,7 @@
         <ImageButton
             android:id="@+id/deleteItemButton"
             android:layout_width="50dp"
+            android:onClick="deleteCartItem"
             android:layout_height="match_parent"
             android:background="@drawable/activity_cart_unselectall_button_rounded"
             android:paddingHorizontal="15dp"
